Generate Main method (#5997)
Make the Jit available from the ILCompiler.ReadyToRun project. I have a PR to refactor the Jit into a separate project in master but didn't want to block on that.

Restrict the root providers in ready-to-run mode such that the entry point is the only thing that acts as a root (other than the ready-to-run header).

Add DebugInformationProvider and DevirtualizationManager to `ReadyToRunCodegenCompilationBuilder`. I thought we could simplify and remove them but that was wrong. Using a combination of EmptyMetadataManager, NullDebugInformationProvider, and the default implementation of DevirtualizationManager is sufficient to compile Main.