Partial implementation of ReadyToRun header tables
This change implements most of the algorithms used for encoding
ReadyToRun header tables and adds logic for populating some of them.
While the logic is by no means complete, it seems to be the minimum
required to make R2RDump not crash while dumping the R2R PE file.

I haven't yet been able to fully execute the PE file in CoreRun
- it crashes while loading the Program type for the Main method;
I suspect there is something wrong with the type table but
I haven't yet been able to debug to the bottom of this.

I'll certainly welcome both semantic and organizational feedback
- I might have misunderstood some of the data or code relationships;
also, some stuff should be probably placed elsewhere (ByteArrayComparer?)
or some of the logic might merit cleaning up. For now I'm trying to
push ahead with producing code so that we can subsequently parallelize
execution bring-up with its further cleanup.

1) Runtime function table - encoding is fully implemented, gets
populated in ObjectWriter; on an anecdotal HelloWorld-type app
the R2R dump shows the Main method in the function table.

2) Method & instance entry point tables - encoding should be fully
implemented, so far it's only been tested for non-generic methods.

3) Types table - encoding should be fully implemented, it is getting
populated in the ObjectWriter. As I mentioned above, it is likely
somewhat wrong as CoreRun fails while loading the Program type.

4) Import sections table - encoding is partially implemented,
currently not populated (just produces an empty table). There is an
open problem regarding population of the section directory entries
for the import sections and I have yet to figure out how to solve it.

Thanks

Tomas