Move the R2R header tables into ReadyToRun subfolder
As JanK pointed out, according to the coding style the folder structure
should match the structured namespace. As I changed the namespace for
these classes from

Compiler.DependencyAnalysis

to

Compiler.DependencyAnalysis.ReadyToRun

I'm also moving the files to a ReadyToRun subfolder to be consistent
with the standard coding style.

Thanks

Tomas