Make emission of managed resources controlled by a policy (#6031)
This makes it possible for the compiler to filter what managed resources should be included in the executable image.

I'm implementing a RESX blocking policy for framework assemblies. This will be used to block emission of framework exception messages. I'm planning to use this from the ProjectX side. Eventually, I would also like to expose this on the CoreRT compiler side. Submitting through git because it's just a better way to do development. I'll follow up on the TFS side and pipe it through NUTC.