Improve bool.TryFormat perf (dotnet/coreclr#18711)
* Improve bool.TryFormat perf

Improves throughput by ~50%, by avoiding AsSpan().CopyTo and just writing out the characters one-by-one.  I experimented with playing the same tricks around storing the data in a ulong and blitting it out to the destination reinterpreted as a span of ulongs, but it didn't make a measurable improvement and increased the code complexity.

* Update Boolean.cs

Signed-off-by: dotnet-bot <dotnet-bot@microsoft.com>