Make the generic composition details RVAs
I'm looking at size regressions between .NET Native 1.7 and the current mainline branch. A thing that stood out is universally bigger EETypes. A contributor to this was getting rid of GenericInstanceDescs between 1.7 and 2.0 and replacing them by direct references to generic composition details from the EEType. Size-wise, this was mostly a wash, but we can actually do better - these new fields are not critical to be pointer-sized.

This change makes turns them into relative pointers. This saves 33 kB on a hello world app. I expect around 100 kB savings on the UWP People app.