CorInfoImpl refactoring part #2
In this part I have mechanically moved parts of CorInfoImpl
to be forked to both CorInfoImpl.RyuJit and CorInfoImpl.ReadyToRun.
I decided not to move getFieldInfo for now as it's a lengthy
method and right now I'm only changing one line in it so that
conditional compilation should be sufficient. We can fork it
later if the changes become more pronounced e.g. due to the
differences in thread static bases.

The full list of methods I'm moving in this change is as follows:

(*) ComputeLookup
(*) getReadyToRunHelper
(*) getReadyToRunDelegateCtorHelper
(*) GetHelperFtnUncached
(*) getFunctionEntryPoint
(*) constructStringLiteral

Apart from the mechanical move I have made a few other minor
changes:

1) I have changed the hard-coded CORINFO_RUNTIME_ABI to a
symbolic constant that is defined in CII.RJ and CII.RTR. In
this one simple case I have already "executed" the split
by putting the appropriate constant into each file, otherwise
I left the moved methods unchanged for now.

2) I have made one additional change I overlooked in my
previous cleanup change: I modified getHelperFtn by adding
optional support for indirection cells that are used for
the helpers in R2R.

As mentioned above, for now I assume that I'll deal with the
differences in getCallInfo and getFieldInfo using

Thanks

Tomas