WIP: Remaining CPAOT changes rebased to my latest changes
This change represents the remainder of my CoreRT R2R bring-up changes
from the last two weeks. Amy, Simon and myself managed to gradually
siphon parts of the complex change off to separate commits so that the
remaining bits have become easier to manage and review.

The only way to further split this remaining change, if desired, would
be to basically by replicacing the bring-up process - enabling just a
subset of the import cells and taking the appropriate parts of the
RTRNF / CII.RTR changes; but I don't think the change is that complex
as is to merit what would otherwise be just busywork.

Feel free to leave out the map file emission if you want to even though
I find it quite useful myself. I want to improve it in the future by
making it a full-fledged map file - storing the entries until the
final relocation phase and then outputting them in section order and
with the full RVA's, that would be much easier for looking up the
various indirection cells. Of course we should probably produce the file
under some switch, I haven't wired that up yet as I'm using the map
all the time.

To expedite the review process, I'm sending this change out for PR
as WIP even though I most likely won't be able to finish the review
process. The main purpose of this review is to solicit early feedback
from Jan and Michal.

One potentially problematic thing is the modification to
MultiFileCompilationModuleGroup. Michal and / or Jan should say whether
we need to derive a new ReadyToRunCompilationModuleGroup and move this
functionality there.

The GC info is still unfinished - in particular, I don't know how to
get to the unwind info. IIRC the code today emits something like an
ulong(0) which should mean "no unwind info" before the GC info. This
is something we'll need to look into soon to unblock larger-scale
testing - not necessarily saying it needs fixing in this change.

As for the conditional directives in getCallInfo / getFieldInfo,
it seems to me that they use just about 2-3 different constructs so
that we might be able to further simplify this by creating
forked helpers for them in the two files CII.RJ / CII.RTR and
remove the conditional compilations.

I leave it up to Simon / Amy to decide whether they want to finish
the review I started by pushing additional commits into my remote
GIT repository and by removing the WIP status or whether they'll
start their own review (or series of reviews) in case of more
substantial modifications to my changes.

In its current form the change also still includes the change to the
repro Program.cs making it the "R2R unit test suite". This shouldn't
be merged in to the repro location. We need to figure out where
to put it, I would definitely love to not throw it away, certainly
not before we have some other testing up and running. For bringing
up the individual helpers it's just awesome.

Thanks

Tomas