Initial support for multi-module compilation
This change basically introduces a new helper struct ModuleToken
encapsulating the combination of an EcmaModule context module
and a token relative to that module. This combo allows for full
token resolution in all cases. Most of the change basically
mechanically replaces mdToken with the new ModuleToken type.

In R2RCGCompilation, we construct CorInfoImpl instances lazily
per EcmaModule; this allows for back-propagation of module info
to R2RCGNF.