Only restore CoreCLR when tests run
Instead of restoring CoreCLR during a build, which is wasteful if `build.cmd skiptests` is called during CI, restore when the test build runs. As a developer aid, runtest.cmd checks if CoreCLR is present and invokes the test-build if necessary. This covers cases where we typically `build.cmd skiptests` and later invoke `runtest.cmd` manually.