Fix test CI failures
Debug mode was failing because the unit test relied on a specific file included in client Windows OS installations that doesn't exist on server OS SKUs. This file won't exist on Unix either, so switch to a text file in the test folder and add support in the harness for test-specific arguments.

Disable release mode for the test. I disabled the scanner for ready-to-run (which was causing VTable generating to include method entry points for System.Object's virtuals) then hit an exception because the Throw ready-to-run helper isn't implemented yet.