Addressing PR feedback
1) Per Michal's suggestion, I have created a new property
OffsetOfDelegateFirstTarget with separate definitions in CII.RyuJit
vs. CII.ReadyToRun. Please note this is somewhat different than
TargetABI because TargetABI is a plain constant whereas the offset
is a function of the target pointer size.

2) I have implemented the IEquatable interface on all three helper
structs used for dictionary indexation in R2RCGNF.

3) Per Simon's suggestion I have moved the non-reloc dependency
in DelegateCtorSignature to the new override
ComputeNonRelocationBasedDependencies. I have used the same technique
in MethodWithGCInfo where it substantially simplifed the code.

Thanks

Tomas