Initial support for instantiating stubs
This change mostly involves propagation of the new instantiating stub
flag around method signature generation. I'm still not fully clear on
when exactly we should ask for the instantiating stub to be generated;
for now it's used for the generic method dictionary cell - this change
(and changing the method dictionary cell to a precode cell) fixes
the two Array.Empty<string> tests in the trivial R2R unit test suite.

Thanks

Tomas