Additional delta to make basic try - catch scenario work
The biggest change is that I removed all explicit passing of
tokens that was causing tons of noisy changes when in practice
we don't even receive the right tokens.

As of this iteration, all type, method and field lookup is based
on pre-recorded information from resolveToken which should be
after all the "ultimate source of truth" for JIT itself.

Follows a detailed description of other changes of note:

ReadyToRunObjectWriter:

Encoding EH info was tricky as CorInfoImpl provides it in form
of ObjectNode.ObjectData instances. After trying out several
variants I added a new feature to ObjectWriter so that some
ObjectNode's can also act as containers of other ObjectData items.

One other change to that file relates to the need to populate the
ExceptionData PE header directory structure.

DelayLoadHelperThunk:

I have renamed DelayLoadHelperThunk to ImportThunk and I modified
it so that it can be also used for eager thunks.

MethodGCInfoNode, MethodWithGCInfo:

After days of labor I have finally figured out how to properly
encode the unwind and GC info on methods so that the runtime
can deal with it properly.

ReadyToRunHelperSignature:

This is the bug uncovered by my R2RDump improvements - we were
using incorrect encoding for the helpers and fixups - these
should use the compressed uint encoding, not single byte
encoding.

RuntimeFunctionsTableNode:

One point of note is the use of a local field "TableSize"
which gets populated when placing (upon calling GetData with
relocsOnly = false) and subsequenly used to populate the
ExceptionData directory entry in the PE header.

It seems to me like wasteful busywork to lay out the runtime
function table a second time just to measure its size. If we don't
like this, we could expand the SymbolTarget to record the end
of the ObjectData record that defined the symbol.

CorInfoImpl:

Adds implementation for CoreCLR encoding of EH info which is
different from the CoreRT EH info encoding.

Thanks

Tomas