Simple R2R fixes in NativeFormatWriter and R2R helper signature
This is the first part of my recent changes, hopefully comprising
non-controversial straightforward bug fixes.

In NativeFormatWriter, this is hopefully the last piece I missed
in my original rewrite; it started failing once we enabled testing
larger test suites with more functions.

In the helper signature, the fiuxp and helper need to use the
signature compressed uint encoding, not plain byte formatting
(I hit this with the personality routine which has helper code
0xF0 with 2-byte compressed encoding.

Thanks

Tomas