WIP: Latest batch of my CPAOT changes resolved against latest
This is the squashed set of my remaining changes rebased against
the latest CoreRT R2R state. I believe the changes fall into the
following categories:

1) Generalization of DelayLoadHelperThunk to also work for eager
thunk - in light of this fact I renamed it to just ImportThunk.

2) Scattered fixes in EH and unwind info to light up exception
handling. This includes the changes to cater for the fact that
CoreCLR uses a different encoding for the EH info than CoreRT.

3) Experimental refactoring to get rid of explicit token passing
so that they are always synthesized based on pre-recorded
resolveToken information.

My current thinking is that I'm going to send out a separate PR
for the ImportThunk first as that's a very straightforward change.
After that I'll try to separate out the EH / unwind info fixes
from the token refactoring and also send them out separately.

Thanks

Tomas