Fix exception handling info in R2R images
1) ReadyToRunObjectWriter support for "nested ObjectData" where a given node
can produce an enumeration of subordinate ObjectData items. This is used
to produce the EH info that comes from CorInfoImpl in form of ObjectData
blobs.

2) One other change in ReadyToRunObjectWriter is the newly added support
for modifying the PE header directories - this is needed to publish
the RUNTIME_FUNCTIONS R2R header table in the Exception Directory,
otherwise RtlLookupFunctionEntry doesn't work.

3) Fix the EH info itself in CorInfoImpl as CoreCLR runtime that we're
targeting with the R2R images uses a different encoding than CoreRT.

4) MethodGCInfoNode is created for each funclet and holds the index
of the corresponding FrameInfo which is used to emit the proper unwind
and GC info for each funclet.

5) RuntimeFunctionsTableNode - emit a separate RUNTIME_FUNCTION entry
for each funclet and emit the appropriate final sentinel entry.

6) ReadyToRunCodegenNodeFactory - new helpers, support for personality
routines, rooting exception lookup table.

Thanks

Tomas