Additional cleanup based on Michal's feedback
I have refactored / simplified emission of the exception handling
info. The crucial observation is that, in CoreCLR, the EH info
is a dumb blob that doesn't contain any relocs. We can leverage
this information to simplify the emitter.

I have removed the logic for emitting subordinate ObjectData
nodes from the object writer and from ExceptionInfoLookupTableNode.
I have deleted SyntheticSymbolNode and instead I introduced
a primitive EHInfoNode that gets populated from the lookup table
in LayoutMethodsWithGCInfo.

In ReadyToRunObjectWriter I kept the previous move of object
emission data to the dedicated method EmitObjectData as I believe
it makes the method EmitPortableExecutable easier to read.

I found out MethodEHInfoNode was not used for anything - it was
a leftover from one of my earlier failed attempts to attack
EH info - so I deleted it.

In CorInfoImpl, I moved the EH info emitters into the forked files
CorInfoImpl.RyuJit / ReadyToRun as the EH info is completely
different for the two targeting runtimes.

Thanks

Tomas