Fix exception handling info in R2R images (#6321)
* Fix exception handling info in R2R images

1) ReadyToRunObjectWriter support for "nested ObjectData" where a given node
can produce an enumeration of subordinate ObjectData items. This is used
to produce the EH info that comes from CorInfoImpl in form of ObjectData
blobs.

2) One other change in ReadyToRunObjectWriter is the newly added support
for modifying the PE header directories - this is needed to publish
the RUNTIME_FUNCTIONS R2R header table in the Exception Directory,
otherwise RtlLookupFunctionEntry doesn't work.

3) Fix the EH info itself in CorInfoImpl as CoreCLR runtime that we're
targeting with the R2R images uses a different encoding than CoreRT.

4) MethodGCInfoNode is created for each funclet and holds the index
of the corresponding FrameInfo which is used to emit the proper unwind
and GC info for each funclet.

5) RuntimeFunctionsTableNode - emit a separate RUNTIME_FUNCTION entry
for each funclet and emit the appropriate final sentinel entry.

6) ReadyToRunCodegenNodeFactory - new helpers, support for personality
routines, rooting exception lookup table.

* Deleted a bit of previously overlooked dead code from R2R object writer

* Minor extra cleanup in the R2R object writer

With the proposed changes, we no longer need to pass around
the SectionBuilder explicitly as we have it in a field.

* Additional cleanup based on Michal's feedback

I have refactored / simplified emission of the exception handling
info. The crucial observation is that, in CoreCLR, the EH info
is a dumb blob that doesn't contain any relocs. We can leverage
this information to simplify the emitter.

I have removed the logic for emitting subordinate ObjectData
nodes from the object writer and from ExceptionInfoLookupTableNode.
I have deleted SyntheticSymbolNode and instead I introduced
a primitive EHInfoNode that gets populated from the lookup table
in LayoutMethodsWithGCInfo.

In ReadyToRunObjectWriter I kept the previous move of object
emission data to the dedicated method EmitObjectData as I believe
it makes the method EmitPortableExecutable easier to read.

I found out MethodEHInfoNode was not used for anything - it was
a leftover from one of my earlier failed attempts to attack
EH info - so I deleted it.

In CorInfoImpl, I moved the EH info emitters into the forked files
CorInfoImpl.RyuJit / ReadyToRun as the EH info is completely
different for the two targeting runtimes.

* Additional MethodGCInfoNode cleanup

In this additional delta I simplified MethodGCInfoNode management.
I changed the logic so that there's just one MethodGCInfoNode per
method; apart from requiring fewer allocations, this let me
drop the delayed factory logic in MethodWithGCInfo.

* Removed obsolete comment

* Addressing Simon's PR feedback