Synthetic token manager
This is the last remainder of my recent CPAOT changes. I'm proposing
to modify the default mechanism for providing tokens to the R2R
compiler to use pre-recorded tokens from calls to resolveToken
as passing them around is cumbersome and sometimes error prone.

As an exception to this new model I'm still passing around tokens
for string literals because I believe that, in this particular case,
it is much faster than constructing a secondary string token
dictionary.

Michal pointed out that it's yet unclear whether some parts of the
CoreCLR runtime may not require referring to some entity via a
particular token, not just an arbitrary pre-recorded one.

For now I'm not aware of such a place; my thinking is that, if
some exist, we can start selectively propagate the token in that
particular case and hopefully keep the default logic in most
places as it's much easier to use and less divergent from RyuJit.

Thanks

Tomas