Generate debug info for r2r images (#6322)
* Generate debug info for r2r images

Build the debug info table when generating a ready-to-run binary. This contains variable / and native to IL offset information.

The JIT calls setVars / setBoundaries to provide both sets of information. These methods are now split for RyuJit / ReadyToRun compilation modes since we use the data differently in each. For instance, in RyuJit mode, we map the native code back to sequence points and directly back to line numbers. In ReadyToRun, we use the raw debug information.

The types `DebugVarInfo` and `DebugLocInfo` are used to transfer this information. I adjusted them slightly to allow them to carry debug info for both RyuJit and ReadyToRun. We might be able to factor out something cleaner if we have a different interface specification for `IMethodCodeNode` in each project.

The debugging data structures are different in RyuJit vs ready-to-run. Replace `IMethodCodeNode` with `IReadyToRunMethodCodeNode` for ready-to-run compilation and split their uses in `CorInfoImpl`.

Factor a bunch of the RyuJit-specific debug bookkeeping out into `CorInfoImpl.RyuJit`

Move `VarLocType` to live next to `VarLoc` and add an accessor to the `VarLoc` struct for returning the location type.