Compile against CoreCLR framework
In ready-to-run mode, use the CoreCLR shared framework implementation
assemblies to generate test R2R images instead of the AOT versions
currently used in standard RyuJit mode.

System.Object has a pointer field for the EEType in each object header
in the AOT implementation of System.Private.CoreLib. The CoreCLR
System.Object does not have this field, and instead the runtime adds an
implicit pointer field for the MethodTable. Handle this in the R2R
compiler with a custom layout algorithm for System.Object.

Adjust the test scripting to consume the CoreCLR framework.
Assert if we accidentally pass in the AOT System.Private.CoreLib in R2R
mode.

CoreCLR all tests pass rate improved by 21 tests to 1370/2347 passing
(58.4%).