Gracefully handle non-compilable methods (down to 336 failures total)
While I guess this additional change may be somewhat controversial,
it has a huge effect on the CoreCLR all tests, bumping up pass rate
to 85% and leaving just 5 remaining ILC failures; so I'd be grateful
for help in figuring out how to finalize the change so that it can be
merged in.

What basically happens is that today, any time a method fails to
compile, it will tear down the entire compilation. This is because
we never PublishCode so that the MethodWithGCInfo forever returns
"StaticDependenciesAreComputed => false" and that hits an assertion
failure in the dependency graph walker.

The proposed change gracefully handles this situation by publishing
an "empty method code"; such methods are not put into the runtime
function or entry point table however they formally satisfy enough
functionality to let the compilation proceed.

Thanks

Tomas