Non-GC TLS fixes
This change fixes a few remaining issues I found in the support
for GC/non-GC in the CPAOT compiler.

1) We need to override the ComputeHasGCStaticBase method in
ReadyToRunCompilerContext as the version in CompilerTypeSystemContext
has the CoreRT-specific check that puts all TLS under the GC base.

2) In ReadyToRunMetadataFieldLayoutAlgorithm, we need a similar
adjustment for the TLS offset like we have for normal non-GC statics;
namely we need to mimic CoreCLR runtime behavior where the statics
blob is introduced by a control structure (DomainLocalModule /
ThreadLocalModule) and furthermore shifted by the number of DefType
records in the MSIL module in accordance with

https://github.com/dotnet/coreclr/blob/99bf5ce15efc1179b47d1e1d1522673a30ae8628/src/vm/ceeload.cpp#L2026

I have added a new base ReadyToRunUnit test that was failing before
and is passing with my change.

Thanks

Tomas