Avoid crashing the compilation on bad InternalCall (#6365)
CoClasses have an `extern`/`InternalCall` constructor that the compiler is choking on because we would hit it in a place where we can't recover. With the check removed, we're going to emit a warning and generate a throwing method body for the constructor.

This is really just about taking the `extern` without `InternalCall` code path for this.