Targeted fixes for CoreCLR test compilation & execution
This is a subset of the modifications I presented in the long-running
WIP PR. I believe these to be well-defined and hopefully non-controversial.
The change basically improves synthetic token harvesting and fixes
a bunch of corner cases I identified in local debugging of the CoreCLR
tests using the CPAOT compiler. My initial assumption is that most of
the changes should be self-explanatory, please let me know if your
perception is different so that I can improve naming, comments or whatever.

Thanks

Tomas