Make dependency analysis viewer work with CoreRT
I always used a local hack to make the dependency analysis viewer tool work. The tool was crashing the compilation because with the viewer attached we started asking for node names before the compilation unit prefix was set.

At some point we should make CompilationUnitPrefix immutable like everything else to prevent this sort of bugs.